+++ CloudShell.ResourceDefinitions.ReferenceProviders/ConfigurationStore/ConfigurationStoreResourceTypeProvider.cs
@@ -15,7 +15,7 @@ public sealed class ConfigurationStoreResourceTypeProvider :
     {
         public static readonly ResourceAttributeId ConfigurationKind = "configuration.kind";
         public static readonly ResourceAttributeId Endpoint = "configuration.endpoint";
-        public static readonly ResourceAttributeId EntryCount = "configuration.entries";
+        public static readonly ResourceAttributeId EntryCount = "configuration.entries.count";
     }
 
     public static class Operations